GOR-R: R library for remote execution of GOR queries and authorization with WuXi NextCode gateway API.

GOR-R: R library for remote execution of GOR queries and authorization with WuXi NextCode gateway API.

GOR-R functions

The main functions you'll be using are:

gor_connect to establish a connection to your GOR API, and

gor_query to run your queries

Examples

# NOT RUN {
library(gorr)
# Go to your CSA instance of choice that runs the Query API, e.g.
# https://4-20-test.wuxinextcodedev.com
# Note the internal names of the projects you have access to
# go to and copy your api_key from: https://4-20-test.wuxinextcodedev.com/api-key-service/token
# Paste it into the variable below:

api_key <- ""

# Make connection object
conn <- gor_connect(api_key, "test_proj")

# Print the connection details
print(conn)

# Run a simple query

results <- gor_query("gor #dbsnp# | top 10000", conn)

print(results)
# }