Upload phenotype data

The data is expected to be a data.frame, tibble or list of lists. If data.frame or tibble is provided it should contain 2 columns. and list of lists should be of the form list(list(pn1, value1),list(pn2,value2)) or list(list(pn1),list(pn2)) For phenotypes of result_type SET. The result_type of the phenotype dictates if each sublist should contain one or two items.

phenotype_upload_data(phenotype, data, conn = NULL)

Arguments

phenotype

phenotype structure, create or get it using create_phenotype or get_phenotype

data

a data.frame or list of lists to be uploaded

For phenotypes of result_type 'SET' - data should be either a list single column data.frame.

For other phenotypes the data.frame/tibble should contain 2 columns, pn and value or if list a pn value pair for each pn.

conn

Deprecated : gor connection structure, create it using platform_connect